fix done user, device,landing page

+++ src/Controller/UsersController.php
@@ -2,6 +2,11 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Datasource\ConnectionManager;
+use App\Model\Entity\User;
+use Cake\Event\Event;
+use Cake\ORM\TableRegistry;
+use Cake\Utility\Hash;
 
 /**
  * Users Controller
@@ -20,13 +25,15 @@ class UsersController extends AppController
      */
     public function index()
     {
-        $this->paginate = [
-            'contain' => ['Adgroups', 'Landingpages', 'Devices', 'ServiceGroups']
-        ];
-        $users = $this->paginate($this->Users);
-
+        $users = $this->Users->find('all',[
+            'contain'  => ['Landingpages' => function ($q) {
+                return $q
+                    ->where([
+                        'Users.delete_flag !=' => 1
+                    ]);
+            }]
+        ])->toArray();
         $this->set(compact('users'));
-        $this->set('_serialize', ['users']);
     }
 
     /**
@@ -39,7 +46,7 @@ class UsersController extends AppController
     public function view($id = null)
     {
         $user = $this->Users->get($id, [
-            'contain' => ['Adgroups', 'Landingpages', 'Devices', 'ServiceGroups']
+            'contain' => ['Landingpages']
         ]);
 
         $this->set('user', $user);
@@ -53,21 +60,30 @@ class UsersController extends AppController
      */
     public function add()
     {
+        $this->getAllData();
+        $conn = ConnectionManager::get('default');
+        $conn->begin();
         $user = $this->Users->newEntity();
         if ($this->request->is('post')) {
             $user = $this->Users->patchEntity($user, $this->request->getData());
-            if ($this->Users->save($user)) {
-                $this->Flash->success(__('The user has been saved.'));
-
-                return $this->redirect(['action' => 'index']);
+            $user->delete_flag = '0';
+            if (empty($user->errors())) {
+                if ($this->Users->save($user)) {
+                    $conn->commit();
+                    $this->Flash->success(__('The user has been saved.'));
+                    return $this->redirect(['action' => 'index']);
+                } else {
+                    $conn->rollback();
+                    $this->Flash->error(__('The user could not be saved. Please, try again.'));
+                    return $this->redirect(['action' => 'add']);
+                }
+            } else {
+                $conn->rollback();
+                $this->Flash->error(__('The user could not be saved. Please, try again.'));
+                return $this->redirect(['action' => 'add']);
             }
-            $this->Flash->error(__('The user could not be saved. Please, try again.'));
         }
-        $adgroups = $this->Users->Adgroups->find('list', ['limit' => 200]);
-        $landingpages = $this->Users->Landingpages->find('list', ['limit' => 200]);
-        $devices = $this->Users->Devices->find('list', ['limit' => 200]);
-        $serviceGroups = $this->Users->ServiceGroups->find('list', ['limit' => 200]);
-        $this->set(compact('user', 'adgroups', 'landingpages', 'devices', 'serviceGroups'));
+        $this->set(compact('user'));
         $this->set('_serialize', ['user']);
     }
 
@@ -78,46 +94,68 @@ class UsersController extends AppController
      * @return \Cake\Http\Response|null Redirects on successful edit, renders view otherwise.
      * @throws \Cake\Network\Exception\NotFoundException When record not found.
      */
+
     public function edit($id = null)
     {
+        $conn = ConnectionManager::get('default');
+        $conn->begin();
+        $this->getAllData();
+        if (!$this->Landingpages->exists($id)) {
+            $this->redirect(array('controller' => 'Users', 'action' => 'index'));
+        }
+
         $user = $this->Users->get($id, [
             'contain' => []
         ]);
         if ($this->request->is(['patch', 'post', 'put'])) {
             $user = $this->Users->patchEntity($user, $this->request->getData());
-            if ($this->Users->save($user)) {
-                $this->Flash->success(__('The user has been saved.'));
-
-                return $this->redirect(['action' => 'index']);
+            if (empty($user->errors())) {
+                if ($this->Users->save($user)) {
+                    $this->Flash->success(__('The user has been saved.'));
+                    return $this->redirect(['action' => 'index']);
+                } else {
+                    $conn->rollback();
+                    $this->Flash->error(__('The user could not be saved. Please, try again.'));
+                    return $this->redirect(['action' => 'edit'.'/'. $id]);
+                }
+            } else {
+                $conn->rollback();
+                $this->Flash->error(__('The user could not be saved. Please, try again.'));
+                return $this->redirect(['action' => 'edit'.'/'. $id]);
             }
-            $this->Flash->error(__('The user could not be saved. Please, try again.'));
         }
-        $adgroups = $this->Users->Adgroups->find('list', ['limit' => 200]);
-        $landingpages = $this->Users->Landingpages->find('list', ['limit' => 200]);
-        $devices = $this->Users->Devices->find('list', ['limit' => 200]);
-        $serviceGroups = $this->Users->ServiceGroups->find('list', ['limit' => 200]);
-        $this->set(compact('user', 'adgroups', 'landingpages', 'devices', 'serviceGroups'));
+        $this->set(compact('user'));
         $this->set('_serialize', ['user']);
     }
 
     /**
      * Delete method
      *
-     * @param string|null $id User id.
      * @return \Cake\Http\Response|null Redirects to index.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function delete($id = null)
+    public function delete()
     {
+        $this->autoRender = false;
+        $conn = ConnectionManager::get('default');
+        $conn->begin();
         $this->request->allowMethod(['post', 'delete']);
-        $user = $this->Users->get($id);
-        if ($this->Users->delete($user)) {
-            $this->Flash->success(__('The user has been deleted.'));
-        } else {
-            $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+        if ($this->request->getData()) {
+            $user = $this->Users->get($this->request->getData('id'));
+            $user->delete_flag = true;
+            if (empty($user->errors())) {
+                if ($this->Users->save($user)) {
+                    $conn->commit();
+                    die(json_encode(true));
+                } else {
+                    $conn->rollback();
+                    die(json_encode(false));
+                }
+            } else {
+                $conn->rollback();
+                die(json_encode(false));
+            }
         }
-
-        return $this->redirect(['action' => 'index']);
     }
 	
 	/**
@@ -144,19 +182,23 @@ class UsersController extends AppController
             if (empty($user->errors())) {
                 $user = $this->Auth->identify();
                 if ($user) {
-                    $keep_login = isset($this->request->getData()['keep_login']) ? $this->request->getData()['keep_login']:'';
-                    if ($keep_login) {
-                        $session = array(
-                            'id' => $user['id'],
-                            'email' => $this->request->getData()['email'],
-                            'password' => $this->request->getData()['password'],
-                        );
-                        $this->request->session()->write('Users', $session);
+                    if ($user['delete_flag'] != DELETED) {
+                        $keep_login = isset($this->request->getData()['keep_login']) ? $this->request->getData()['keep_login']:'';
+                        if ($keep_login) {
+                            $session = array(
+                                'id' => $user['id'],
+                                'email' => $this->request->getData()['email'],
+                                'password' => $this->request->getData()['password'],
+                            );
+                            $this->request->session()->write('Users', $session);
+                        } else {
+                            $this->request->session()->delete('Users');
+                        }
+                        $this->Auth->setUser($user);
+                        return $this->redirect($this->Auth->redirectUrl());
                     } else {
-                        $this->request->session()->delete('Users');
+                        $this->Flash->error(__(ERROR_USERS_OR_PASSWORD));
                     }
-                    $this->Auth->setUser($user);
-                    return $this->redirect($this->Auth->redirectUrl());
                 } else {
                     $this->Flash->error(__(ERROR_USERS_OR_PASSWORD));
                 }
@@ -175,6 +217,63 @@ class UsersController extends AppController
      */
     public function logout()
     {
+        $this->request->session()->delete('Users');
         return $this->redirect($this->Auth->logout());
+    }
+
+    /**
+     * isEmailExist method
+     * check email when add new users
+     * @return \Cake\Http\Response|null Redirects to index.
+     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     */
+    public function isEmailExist()
+    {
+        $this->autoRender = false;
+        if ($this->request->getData()) {
+            $query = $this->Users->find('all', array(
+                    'recursive' => -1,
+                    'conditions' => array(
+                        'email' => $this->request->getData('email'),
+                        'delete_flag !=' => DELETED
+                    ))
+            );
+            $number = $query->count();
+            if (!$number) {
+                die(json_encode(true));
+            } else {
+                die(json_encode(false));
+            }
+        }
+    }
+
+    /**
+     * isEmaiEditlExist method
+     * check email when edit new users
+     * @return \Cake\Http\Response|null Redirects to index.
+     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     */
+    public function isEmaiEditlExist()
+    {
+        $this->autoRender = false;
+        if ($this->request->getData()) {
+            if ($this->request->getData('email') !== $this->request->getData('backup_email')) {
+                $query = $this->Users->find('all', array(
+                        'recursive' => -1,
+                        'conditions' => array(
+                            'email' => $this->request->getData('email'),
+                            'delete_flag !=' => DELETED
+                        ))
+                );
+                $number = $query->count();
+                if (!$number) {
+                    die(json_encode(true));
+                } else {
+                    die(json_encode(false));
+                }
+            } else {
+                die(json_encode(true));
+            }
+        }
     }
 }
